+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 
 - Added missing functions in controllers
 - Finished camera router apart from finding cameras by address
+- Started writing tests for the parking spots router